RT-75: вынес RandString в utils
